Move map segment stitching into neutral cartographyPathUtils.
Stops mapCellUnionBoundary from depending on nebulaCloudOverlay for generic geometry.
